+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 `StretchScrollView` adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0](https://github.com/APUtils/StretchScrollView/releases/tag/1.2.0)
+Released on 08/04/2017.
+
+#### Changed
+- Migrated to APExtensions 3.3.0
+- Updated GIF
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
 ## [1.1.0](https://github.com/APUtils/StretchScrollView/releases/tag/1.1.0)
 Released on 07/20/2017.